feat: enrich trace payload capture, redaction, and inspection
Backend:
- Add DebugTracePayloadKind enum (INPUT/OUTPUT/TOOL_ARGS/TOOL_RESULT/ERROR)
  and normalize legacy payload_kind values via migration.
- Redact secrets and hidden reasoning before persisting debug payloads
  (DebugTraceContentRedactor).
- Capture structured step I/O (routing/direct_answer/review/drawing) against
  telemetry step spans; keep run output + final diagramId together.
- Replace the runId-keyed streamOutputCaptures map with a request-local
  BoundedTextCapture (removes a latent leak; idempotent, thread-safe).
- Track LLM provider request/response ids, TTFT, attempt/retry counts;
  widen original_length to BIGINT; add changed_cell_count snapshot diff.

Fix:
- Direct answers can resolve to null; Map.of(...) NPE'd inside the step
  capture and wrongly flipped a successful run into a FAILED span (and threw
  to the client). Use a null-tolerant traceField() for step payload maps.
  Add shouldNotFailWhenDirectAnswerResolvesToNull regression test.

Frontend:
- Trace workbench + payload panel: rendered/raw views, search, fullscreen,
  message extraction, redacted/expired/truncated status.

+++ ai-agent-draw-io/ai-agent-draw-io-api/src/main/java/org/zipp/ai/api/dto/AdminDebugTraceCaptureDTO.java
@@ -17,7 +17,7 @@ public class AdminDebugTraceCaptureDTO {
     private String contentType;
     private String content;
     private String contentSha256;
-    private Integer originalLength;
+    private Long originalLength;
     private boolean truncated;
     private Instant contentExpiresAt;
     private Instant contentDeletedAt;
