+++ ai-agent-draw-io/ai-agent-draw-io-domain/src/main/java/org/zipp/ai/domain/agent/model/valobj/debugtrace/DebugTraceCapture.java
@@ -20,7 +20,7 @@ public class DebugTraceCapture {
     private String contentType;
     private String content;
     private String contentSha256;
-    private Integer originalLength;
+    private Long originalLength;
     private boolean truncated;
     private Instant contentExpiresAt;
     private Instant contentDeletedAt;
