+++ ai-agent-draw-io/ai-agent-draw-io-infrastructure/src/main/java/org/zipp/ai/infrastructure/dao/po/DebugTraceCapturePO.java
@@ -18,7 +18,7 @@ public class DebugTraceCapturePO {
     private String contentType;
     private String content;
     private String contentSha256;
-    private Integer originalLength;
+    private Long originalLength;
     private Boolean truncated;
     private Date contentExpiresAt;
     private Date contentDeletedAt;
